Adding test for data generation with a semi-Markov model and improving  test messages.
